+++ src/GLdispatch/vnd-glapi/entry_common.h
