+++ src/GLdispatch/vnd-glapi/entry_x86_64_common.c
@@ -1,7 +1,6 @@
 /*
- * Mesa 3-D graphics library
- *
  * Copyright (C) 2010 LunarG Inc.
+ * Copyright (c) 2015, NVIDIA CORPORATION.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"),
@@ -23,76 +22,58 @@
  *
  * Authors:
  *    Chia-I Wu <olv@lunarg.com>
+ *    Kyle Brenneman <kbrenneman@nvidia.com>
  */
 
 #include "entry.h"
+#include "entry_common.h"
+
+#include <string.h>
+#include <stdint.h>
+#include <sys/mman.h>
+#include <unistd.h>
 #include <assert.h>
-#include <stdlib.h>
 
-#include "glapi/glapi.h"
-#include "glvnd/GLdispatchABI.h"
+#include "glapi.h"
+#include "u_macros.h"
+#include "u_current.h"
+#include "utils_misc.h"
 
-static INLINE const struct _glapi_table *
-entry_current_get(void)
-{
-#if defined (GLDISPATCH_USE_TLS)
-   return _glapi_tls_Current[GLAPI_CURRENT_DISPATCH];
-#else
-   return (likely(_glapi_Current[GLAPI_CURRENT_DISPATCH]) ?
-         _glapi_Current[GLAPI_CURRENT_DISPATCH] : _glapi_get_current());
-#endif
-}
+/**
+ * \file
+ *
+ * Common functions for the x86 and x86-64 stubs.
+ *
+ * These functions are almost identical to the ones used for ARMv7, except that
+ * the ARM stubs have to add 1 to the address of each entrypoint to force
+ * switching to Thumb mode.
+ */
 
-/* C version of the public entries */
-#define MAPI_TMP_DEFINES
-#define MAPI_TMP_PUBLIC_DECLARES
-#define MAPI_TMP_PUBLIC_ENTRIES
-#include "mapi_tmp.h"
-
-const int entry_type = __GLDISPATCH_STUB_UNKNOWN;
-const int entry_stub_size = 0;
-
-void
-entry_init_public(void)
+void entry_init_public(void)
 {
 }
 
-void
-entry_generate_default_code(char *entry, int slot)
+mapi_func entry_get_public(int index)
 {
-    assert(0);
-}
-
-mapi_func
-entry_get_public(int index)
-{
-   /* pubic_entries are defined by MAPI_TMP_PUBLIC_ENTRIES */
-   return public_entries[index];
-}
-
-int entry_patch_start(void)
-{
-    assert(!"This should never be called");
-    return 0;
-}
-
-int entry_patch_finish(void)
-{
-    assert(!"This should never be called");
-    return 0;
+    return (mapi_func)(public_entry_start + (index * entry_stub_size));
 }
 
 void entry_get_patch_addresses(mapi_func entry, void **writePtr, const void **execPtr)
 {
-    assert(!"This should never be called");
-    *writePtr = NULL;
-    *execPtr = NULL;
+    *execPtr = (const void *) entry;
+    *writePtr = u_execmem_get_writable(entry);
 }
 
 #if !defined(STATIC_DISPATCH_ONLY)
-mapi_func
-entry_generate(int slot)
+mapi_func entry_generate(int slot)
 {
-   return NULL;
+    void *code = u_execmem_alloc(entry_stub_size);
+    if (!code) {
+        return NULL;
+    }
+
+    entry_generate_default_code(code, slot);
+
+    return (mapi_func) code;
 }
 #endif // !defined(STATIC_DISPATCH_ONLY)
