+++ src/GLdispatch/vnd-glapi/entry_x86_64_tls.c
@@ -35,7 +35,7 @@
 
 #include "utils_misc.h"
 #include "u_macros.h"
-#include "glapi/glapi.h"
+#include "glapi.h"
 #include "glvnd/GLdispatchABI.h"
 
 #define ENTRY_STUB_ALIGN 32
