+++ src/GLdispatch/vnd-glapi/entry_x86_tsd.c
@@ -35,7 +35,7 @@
 #include <string.h>
 
 #include "u_macros.h"
-#include "glapi/glapi.h"
+#include "glapi.h"
 #include "glvnd/GLdispatchABI.h"
 
 #define X86_ENTRY_SIZE 32
