+++ src/GLdispatch/vnd-glapi/u_current.h
@@ -1,7 +1,7 @@
 #ifndef _U_CURRENT_H_
 #define _U_CURRENT_H_
 
-#include "glapi/glapi.h"
+#include "glapi.h"
 
 /* ugly renames to match glapi.h */
